Refactor Database to extract ConstraintManager
Extracted `ConstraintManager` from `relvar-core/src/database.rs` to a new module `relvar-core/src/constraints/manager.rs`. This reduces the size of the `Database` god struct and improves cohesion by grouping constraint logic with constraint definitions.

Introduced `ConstraintManagerError` to decouple constraint logic from the main `DatabaseError`, while preserving the public API surface via `From` implementation.
