+++ relvar-core/src/constraints/mod.rs
@@ -38,10 +38,12 @@ pub mod check;
 pub mod expression;
 pub mod foreign_key;
 pub mod key;
+pub mod manager;
 pub mod type_constraint;
 
 pub use check::{CheckConstraint, CheckConstraintError, CheckConstraints, CheckPredicate};
 pub use expression::{CmpOp, ConstraintExpression, ExpressionError, ValueOrRef};
 pub use foreign_key::{ForeignKey, ForeignKeyConstraints, ForeignKeyError};
 pub use key::{CandidateKey, KeyConstraintError, KeyConstraints, PrimaryKey};
+pub use manager::{ConstraintManager, ConstraintManagerError};
 pub use type_constraint::{AttributeConstraints, TypeConstraint, TypeConstraintError};
